remove dep maa-dirs for lib
