+++ crates/maa-server/src/lib.rs
@@ -215,209 +215,55 @@ pub mod core {
             let Self {
                 static_ops,
                 log_ops,
+                lib_path,
+                resource_dirs,
             } = self;
 
+            crate::utils::load_core(lib_path.into()).map_err(|e| {
+                tonic::Status::invalid_argument(format!("Failed to load MaaCore due to {e}"))
+            })?;
             if let Some(ops) = log_ops {
                 ops.apply()?;
             }
             if let Some(ops) = static_ops {
                 ops.apply()?;
             }
+            crate::utils::load_resource(resource_dirs).map_err(|e| {
+                tonic::Status::invalid_argument(format!("Failed to load Maa Resource due to {e}"))
+            })?;
             Ok(())
         }
     }
 }
 
-pub mod utils {
-    pub fn load_core() -> Result<(), String> {
-        use maa_dirs::MAA_CORE_LIB;
-        if let Some(lib_dir) = maa_dirs::find_library() {
-            tracing::debug!("Loading MaaCore from: {}", lib_dir.display());
-            // Set DLL directory on Windows
-            #[cfg(target_os = "windows")]
-            {
-                use windows_strings::HSTRING;
-                use windows_sys::Win32::System::LibraryLoader::SetDllDirectoryW;
+mod utils {
+    pub fn load_core(path_to_core: std::path::PathBuf) -> Result<(), String> {
+        tracing::debug!("Loading MaaCore from: {}", path_to_core.display());
+        // Set DLL directory on Windows
+        #[cfg(target_os = "windows")]
+        {
+            use windows_strings::HSTRING;
+            use windows_sys::Win32::System::LibraryLoader::SetDllDirectoryW;
 
-                let code = unsafe { SetDllDirectoryW(HSTRING::from(lib_dir.as_ref()).as_ptr()) };
-                if code == 0 {
-                    return Err(anyhow::Error::new(windows_result::Error::from_win32())
-                        .context("Failed to set DLL directory!"));
-                }
+            let code =
+                unsafe { SetDllDirectoryW(HSTRING::from(path_to_core.parent().unwrap()).as_ptr()) };
+            if code == 0 {
+                return Err(anyhow::Error::new(windows_result::Error::from_win32())
+                    .context("Failed to set DLL directory!"));
             }
-            maa_sys::binding::load(lib_dir.join(MAA_CORE_LIB))
-        } else {
-            tracing::debug!("MaaCore not found, trying to load from system library path");
-            maa_sys::binding::load(MAA_CORE_LIB)
         }
-        .map_err(|e| e.to_string())
+        maa_sys::binding::load(path_to_core).map_err(|e| e.to_string())
     }
 
-    use maa_dirs::{self as dirs, join};
-    use std::path::PathBuf;
-
-    #[cfg_attr(test, derive(Debug, PartialEq))]
-    #[derive(Clone)]
-    pub struct ResourceConfig {
-        /// Resources used by global arknights client, e.g. `YostarEN`
-        global_resource: Option<PathBuf>,
-        /// Resources used by platform diff, subdirectories of `resource_base_dirs`, e.g.
-        /// `platform_diff/iOS`
-        platform_diff_resource: Option<PathBuf>,
-        /// Whether to load resources from user config directory, when enabled, the
-        /// `MAA_CONFIG_DIR/resource` will be appended to `resource_base_dirs` as the last element
-        user_resource: bool,
-        /// Resource base directories, a list of directories containing resource directories
-        /// Not deserialized from config file
-        pub(crate) resource_base_dirs: Vec<PathBuf>,
+    pub fn load_resource(resource_dirs: Vec<String>) -> maa_sys::Result<()> {
+        for resource_dir in resource_dirs {
+            let resource_dir = std::path::PathBuf::from(resource_dir);
+            tracing::debug!("Loading resource from {}", resource_dir.display());
+            maa_sys::Assistant::load_resource(resource_dir.parent().unwrap())?;
+        }
+        Ok(())
     }
 
-    impl Default for ResourceConfig {
-        fn default() -> Self {
-            Self {
-                resource_base_dirs: default_resource_base_dirs(),
-                global_resource: None,
-                platform_diff_resource: None,
-                user_resource: false,
-            }
-        }
-    }
-
-    fn default_resource_base_dirs() -> Vec<PathBuf> {
-        let mut resource_dirs = Vec::new();
-
-        if let Some(resource_dir) = dirs::find_resource() {
-            tracing::debug!("Found resource directory: {}", resource_dir.display());
-            resource_dirs.push(resource_dir.into_owned());
-        } else {
-            tracing::warn!("Resource directory not found!")
-        }
-
-        let hot_update_dir = dirs::hot_update();
-        if hot_update_dir.exists() {
-            tracing::debug!(
-                "Found hot update resource directory: {}",
-                hot_update_dir.display()
-            );
-            resource_dirs.push(join!(hot_update_dir, "resource"));
-            resource_dirs.push(join!(hot_update_dir, "cache", "resource"));
-        } else {
-            tracing::warn!("Hot update resource directory not found!");
-        }
-
-        resource_dirs
-    }
-
-    impl ResourceConfig {
-        pub fn use_user_resource(&mut self) -> &mut Self {
-            if !self.user_resource {
-                self.user_resource = true;
-                push_user_resource(&mut self.resource_base_dirs);
-            }
-            self
-        }
-
-        pub fn use_global_resource(&mut self, resource: impl Into<PathBuf>) -> &mut Self {
-            match self.global_resource.as_ref() {
-                Some(global_resource) => {
-                    tracing::warn!(
-                        "Global resource {} already set, ignoring {}",
-                        global_resource.display(),
-                        resource.into().display(),
-                    );
-                }
-                None => {
-                    let resource = resource.into();
-                    tracing::info!("Using global resource: {}", resource.display());
-                    self.global_resource = Some(resource);
-                }
-            }
-            self
-        }
-
-        pub fn use_platform_diff_resource(&mut self, resource: impl Into<PathBuf>) -> &mut Self {
-            match self.platform_diff_resource.as_ref() {
-                Some(platform_diff_resource) => {
-                    tracing::warn!(
-                        "Platform diff resource {} already set, ignoring {}",
-                        platform_diff_resource.display(),
-                        resource.into().display(),
-                    );
-                }
-                None => {
-                    // should not push to resource_base_dirs as this is not a base resource directory
-                    let resource = resource.into();
-                    tracing::info!("Using platform diff resource: {}", resource.display());
-                    self.platform_diff_resource = Some(resource);
-                }
-            }
-            self
-        }
-
-        /// Get base resource directories
-        pub fn base_dirs(&self) -> &Vec<PathBuf> {
-            &self.resource_base_dirs
-        }
-
-        /// Get all resource directories, including global and platform diff resources
-        pub fn resource_dirs(&self) -> Vec<PathBuf> {
-            let base_dirs = self.base_dirs();
-            let mut resource_dirs = base_dirs.clone();
-            if let Some(global_resource) = self.global_resource.as_ref() {
-                let global_resource_dir = join!("global", global_resource, "resource");
-                let full_paths = dirs::global_path(base_dirs, global_resource_dir);
-                if full_paths.is_empty() {
-                    tracing::warn!("Global resource {} not found", global_resource.display(),);
-                } else {
-                    resource_dirs.extend(full_paths);
-                }
-            }
-            if let Some(platform_diff_resource) = self.platform_diff_resource.as_ref() {
-                let platform_diff_resource_dir =
-                    join!("platform_diff", platform_diff_resource, "resource");
-                let full_paths = dirs::global_path(base_dirs, platform_diff_resource_dir);
-                if full_paths.is_empty() {
-                    tracing::warn!(
-                        "Platform diff resource {} not found",
-                        platform_diff_resource.display(),
-                    );
-                } else {
-                    resource_dirs.extend(full_paths);
-                }
-            }
-
-            resource_dirs
-        }
-
-        pub fn load(&self) -> Result<(), String> {
-            let resource_dirs = self.resource_dirs();
-            for resource_dir in resource_dirs {
-                tracing::debug!("Loading resource from {}", resource_dir.display());
-                maa_sys::Assistant::load_resource(resource_dir.parent().unwrap())
-                    .map_err(|e| e.to_string())?;
-            }
-
-            Ok(())
-        }
-    }
-
-    fn push_user_resource(resource_dirs: &mut Vec<PathBuf>) -> &mut Vec<PathBuf> {
-        push_resource(resource_dirs, dirs::config().join("resource"))
-    }
-
-    fn push_resource(
-        resource_dirs: &mut Vec<PathBuf>,
-        dir: impl Into<PathBuf>,
-    ) -> &mut Vec<PathBuf> {
-        let dir = dir.into();
-        if dir.exists() {
-            resource_dirs.push(dir);
-        } else {
-            tracing::warn!("Resource directory {} not found, ignoring", dir.display(),);
-        }
-
-        resource_dirs
-    }
 }
 
 pub mod callback;
